Backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
psycopg2-binary==2.9.10
langchain==0.3.18
langchain-google-genai==2.0.9
qdrant-client==1.13.2
pandas==1.4.1
python-dotenv==1.0.1
openai==1.63.0